+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Improves error handling for `firestore:delete` when deleting very large documents.
 - Support batchCreate API in Auth Emulator (#2947).
+- Support emulators:export and import for Auth Emulator (#2955).